Add files via upload
A test of the model using myself. The output of the model is "<tf.Tensor: shape=(), dtype=string, numpy=b'bin   lleeee  iin      iiieee    lleeaaaeeee'>".  My interpretations of each error are: two lips are not alike; way to talk is different; environment is different. Ways to solve: Use larger and wider dataset; compute the lip movement by tracking the lip using a line.